Boot users menu

See merge request Mikhail_Sboev/Hotel_System!95

+++ src/main/java/com/epam/hotel/services/implementations/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package com.epam.hotel.services.implementations;
 
 import com.epam.hotel.domains.User;
+import com.epam.hotel.domains.enums.BlockStatus;
 import com.epam.hotel.domains.enums.Permission;
 import com.epam.hotel.repositories.UserRepository;
 import com.epam.hotel.services.UserService;
@@ -40,6 +41,7 @@ public class UserServiceImpl implements UserService {
     public User save(User user) {
         if (user.getId() == 0) {
             user.setPermission(Permission.USER);
+            user.setBlock(BlockStatus.UNBLOCKED);
             user.setPassword(Encoder.encode(user.getPassword()));
         }
         return userRepository.save(user);
@@ -47,7 +49,12 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteById(Long id) {
-        userRepository.deleteById(id);
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.get().getPermission() == Permission.USER){
+            userRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("You cant delete ADMIN");
+        }
     }
 
     @Override
@@ -57,6 +64,20 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Page<User> getAllUsersPaged(int page, int size) { return userRepository.findAll(PageRequest.of(page - 1, size, Sort.Direction.ASC, "id")); }
+
+    @Override
+    public void changeUserBlockForId(Long id) {
+        User user = userRepository.getOne(id);
+        switch (user.getBlock()) {
+            case BLOCKED:
+                user.setBlock(BlockStatus.UNBLOCKED);
+                break;
+            case UNBLOCKED:
+                user.setBlock(BlockStatus.BLOCKED);
+                break;
+        }
+        userRepository.save(user);
+    }
 
 }
